版本同步

+++ backend/src/main/java/io/metersphere/api/service/ApiModuleService.java
@@ -1,28 +1,30 @@
 package io.metersphere.api.service;
 
 
+import com.alibaba.fastjson.JSON;
 import io.metersphere.api.dto.definition.ApiDefinitionRequest;
 import io.metersphere.api.dto.definition.ApiDefinitionResult;
 import io.metersphere.api.dto.definition.ApiModuleDTO;
 import io.metersphere.api.dto.definition.DragModuleRequest;
-import io.metersphere.base.domain.ApiDefinitionExample;
-import io.metersphere.base.domain.ApiModule;
-import io.metersphere.base.domain.ApiModuleExample;
+import io.metersphere.base.domain.*;
 import io.metersphere.base.mapper.ApiDefinitionMapper;
 import io.metersphere.base.mapper.ApiModuleMapper;
 import io.metersphere.base.mapper.ext.ExtApiDefinitionMapper;
 import io.metersphere.base.mapper.ext.ExtApiModuleMapper;
 import io.metersphere.commons.constants.TestCaseConstants;
 import io.metersphere.commons.exception.MSException;
-
 import io.metersphere.i18n.Translator;
 import io.metersphere.service.NodeTreeService;
+import io.metersphere.service.ProjectService;
+import io.metersphere.track.service.TestPlanApiCaseService;
+import io.metersphere.track.service.TestPlanProjectService;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.ibatis.session.ExecutorType;
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.apache.commons.collections.CollectionUtils;
 
 import javax.annotation.Resource;
 import java.util.*;
@@ -40,6 +42,16 @@ public class ApiModuleService extends NodeTreeService<ApiModuleDTO> {
     private ApiDefinitionMapper apiDefinitionMapper;
     @Resource
     private ExtApiDefinitionMapper extApiDefinitionMapper;
+    @Resource
+    private TestPlanProjectService testPlanProjectService;
+    @Resource
+    private ProjectService projectService;
+    @Resource
+    private TestPlanApiCaseService testPlanApiCaseService;
+    @Resource
+    private ApiTestCaseService apiTestCaseService;
+    @Resource
+    private ApiDefinitionService apiDefinitionService;
 
     @Resource
     SqlSessionFactory sqlSessionFactory;
@@ -58,13 +70,84 @@ public class ApiModuleService extends NodeTreeService<ApiModuleDTO> {
         return addNodeWithoutValidate(node);
     }
 
+    private double getNextLevelPos(String projectId, int level, String parentId) {
+        List<ApiModule> list = getPos(projectId, level, parentId, "pos desc");
+        if (!CollectionUtils.isEmpty(list) && list.get(0) != null && list.get(0).getPos() != null) {
+            return list.get(0).getPos() + DEFAULT_POS;
+        } else {
+            return DEFAULT_POS;
+        }
+    }
+
+    private List<ApiModule> getPos(String projectId, int level, String parentId, String order) {
+        ApiModuleExample example = new ApiModuleExample();
+        ApiModuleExample.Criteria criteria = example.createCriteria();
+        criteria.andProjectIdEqualTo(projectId).andLevelEqualTo(level);
+        if (level != 1 && StringUtils.isNotBlank(parentId)) {
+            criteria.andParentIdEqualTo(parentId);
+        }
+        example.setOrderByClause(order);
+        return apiModuleMapper.selectByExample(example);
+    }
+
     public String addNodeWithoutValidate(ApiModule node) {
         node.setCreateTime(System.currentTimeMillis());
         node.setUpdateTime(System.currentTimeMillis());
         node.setId(UUID.randomUUID().toString());
+        double pos = getNextLevelPos(node.getProjectId(), node.getLevel(), node.getParentId());
+        node.setPos(pos);
         apiModuleMapper.insertSelective(node);
         return node.getId();
     }
+
+    public List<ApiModuleDTO> getNodeByPlanId(String planId, String protocol) {
+        List<ApiModuleDTO> list = new ArrayList<>();
+        List<String> projectIds = testPlanProjectService.getProjectIdsByPlanId(planId);
+        projectIds.forEach(id -> {
+            Project project = projectService.getProjectById(id);
+            String name = project.getName();
+            List<ApiModuleDTO> nodeList = getNodeDTO(id, planId, protocol);
+            ApiModuleDTO apiModuleDTO = new ApiModuleDTO();
+            apiModuleDTO.setId(project.getId());
+            apiModuleDTO.setName(name);
+            apiModuleDTO.setLabel(name);
+            apiModuleDTO.setChildren(nodeList);
+            list.add(apiModuleDTO);
+        });
+        return list;
+    }
+
+    private List<ApiModuleDTO> getNodeDTO(String projectId, String planId, String protocol) {
+        List<TestPlanApiCase> apiCases = testPlanApiCaseService.getCasesByPlanId(planId);
+        if (apiCases.isEmpty()) {
+            return null;
+        }
+        List<ApiModuleDTO> testCaseNodes = extApiModuleMapper.getNodeTreeByProjectId(projectId, protocol);
+
+        List<String> caseIds = apiCases.stream()
+                .map(TestPlanApiCase::getApiCaseId)
+                .collect(Collectors.toList());
+
+        List<String> definitionIds = apiTestCaseService.selectCasesBydIds(caseIds).stream()
+                .map(ApiTestCase::getApiDefinitionId)
+                .collect(Collectors.toList());
+
+        List<String> dataNodeIds = apiDefinitionService.selectApiDefinitionBydIds(definitionIds).stream()
+                .map(ApiDefinition::getModuleId)
+                .collect(Collectors.toList());
+
+        List<ApiModuleDTO> nodeTrees = getNodeTrees(testCaseNodes);
+
+        Iterator<ApiModuleDTO> iterator = nodeTrees.iterator();
+        while (iterator.hasNext()) {
+            ApiModuleDTO rootNode = iterator.next();
+            if (pruningTree(rootNode, dataNodeIds)) {
+                iterator.remove();
+            }
+        }
+        return nodeTrees;
+    }
+
 
     public ApiModule getNewModule(String name, String projectId, int level) {
         ApiModule node = new ApiModule();
@@ -87,7 +170,19 @@ public class ApiModuleService extends NodeTreeService<ApiModuleDTO> {
 
     private void checkApiModuleExist(ApiModule node) {
         if (node.getName() != null) {
-            if (selectSameModule(node).size() > 0) {
+            ApiModuleExample example = new ApiModuleExample();
+            ApiModuleExample.Criteria criteria = example.createCriteria();
+            criteria.andNameEqualTo(node.getName())
+                    .andProjectIdEqualTo(node.getProjectId());
+            if (StringUtils.isNotBlank(node.getParentId())) {
+                criteria.andParentIdEqualTo(node.getParentId());
+            } else {
+                criteria.andParentIdIsNull();
+            }
+            if (StringUtils.isNotBlank(node.getId())) {
+                criteria.andIdNotEqualTo(node.getId());
+            }
+            if (apiModuleMapper.selectByExample(example).size() > 0) {
                 MSException.throwException(Translator.get("test_case_module_already_exists") + ": " + node.getName());
             }
         }
@@ -118,21 +213,22 @@ public class ApiModuleService extends NodeTreeService<ApiModuleDTO> {
     public int editNode(DragModuleRequest request) {
         request.setUpdateTime(System.currentTimeMillis());
         checkApiModuleExist(request);
-        List<ApiDefinitionResult> apiModule = queryByModuleIds(request.getNodeIds());
-
-        apiModule.forEach(apiDefinition -> {
-            StringBuilder path = new StringBuilder(apiDefinition.getModulePath());
-            List<String> pathLists = Arrays.asList(path.toString().split("/"));
-            pathLists.set(request.getLevel(), request.getName());
-            path.delete(0, path.length());
-            for (int i = 1; i < pathLists.size(); i++) {
-                path = path.append("/").append(pathLists.get(i));
-            }
-            apiDefinition.setModulePath(path.toString());
-        });
-
-        batchUpdateApiDefinition(apiModule);
-
+        List<ApiDefinitionResult> apiDefinitionResults = queryByModuleIds(request.getNodeIds());
+        if (CollectionUtils.isNotEmpty(apiDefinitionResults)) {
+            apiDefinitionResults.forEach(apiDefinition -> {
+                if (apiDefinition != null && StringUtils.isNotBlank(apiDefinition.getModulePath())) {
+                    StringBuilder path = new StringBuilder(apiDefinition.getModulePath());
+                    List<String> pathLists = Arrays.asList(path.toString().split("/"));
+                    pathLists.set(request.getLevel(), request.getName());
+                    path.delete(0, path.length());
+                    for (int i = 1; i < pathLists.size(); i++) {
+                        path = path.append("/").append(pathLists.get(i));
+                    }
+                    apiDefinition.setModulePath(path.toString());
+                }
+            });
+            batchUpdateApiDefinition(apiDefinitionResults);
+        }
         return apiModuleMapper.updateByPrimaryKeySelective(request);
     }
 
@@ -155,6 +251,18 @@ public class ApiModuleService extends NodeTreeService<ApiModuleDTO> {
         sqlSession.flushStatements();
     }
 
+    @Override
+    public ApiModuleDTO getNode(String id) {
+        ApiModule module = apiModuleMapper.selectByPrimaryKey(id);
+        ApiModuleDTO dto = JSON.parseObject(JSON.toJSONString(module), ApiModuleDTO.class);
+        return dto;
+    }
+
+    @Override
+    public void updatePos(String id, Double pos) {
+        extApiModuleMapper.updatePos(id, pos);
+    }
+
     public void dragNode(DragModuleRequest request) {
 
         checkApiModuleExist(request);
@@ -169,7 +277,7 @@ public class ApiModuleService extends NodeTreeService<ApiModuleDTO> {
         if (nodeTree == null) {
             return;
         }
-        buildUpdateDefinition(nodeTree, apiModule, updateNodes, "/", "0", nodeTree.getLevel());
+        buildUpdateDefinition(nodeTree, apiModule, updateNodes, "/", "0", 1);
 
         updateNodes = updateNodes.stream()
                 .filter(item -> nodeIds.contains(item.getId()))
